--- UACloudLibraryServer/NodeSetIndex/CloudLibDbOpcUaContext.cs
@@ -1,56 +1,0 @@
-/* ========================================================================
- * Copyright (c) 2005-2022 The OPC Foundation, Inc. All rights reserved.
- *
- * OPC Foundation MIT License 1.00
- *
- * Permission is hereby granted, free of charge, to any person
- * obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without
- * restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- *
- * The complete license agreement can be found here:
- * http://opcfoundation.org/License/MIT/1.00/
- * ======================================================================*/
-
-using System;
-using Microsoft.Extensions.Logging;
-using Opc.Ua.Cloud.Library.Models;
-using Opc.Ua.Cloud.Library.NodeSetIndex;
-using Opc.Ua.Export;
-
-namespace Opc.Ua.Cloud.Library
-{
-    internal class CloudLibDbOpcUaContext : DbOpcUaContext
-    {
-        public CloudLibDbOpcUaContext(AppDbContext dbContext, ILogger logger, Func<ModelTableEntry, CloudLibNodeSetModel> nodeSetFactory)
-            : base(dbContext, logger, nodeSetFactory)
-        {
-        }
-
-        public override NodeSetModel GetOrAddNodesetModel(ModelTableEntry model, bool createNew = true)
-        {
-            var nodeSetModel = base.GetOrAddNodesetModel(model) as CloudLibNodeSetModel;
-            if (!createNew && nodeSetModel != null && nodeSetModel.ValidationStatus != ValidationStatus.Indexed)
-            {
-                throw new ArgumentException($"Required NodeSet {nodeSetModel} not indexed yet.");
-            }
-
-            return nodeSetModel;
-        }
-    }
-}
