--- UACloudLibraryServer/NodeSetIndex/NodeModelOpcExtensions.cs
@@ -1,69 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Opc.Ua.Export;
-
-namespace Opc.Ua.Cloud.Library.NodeSetIndex
-{
-    public static class NodeModelOpcExtensions
-    {
-        public static DateTime GetNormalizedPublicationDate(this ModelTableEntry model)
-        {
-            return model.PublicationDateSpecified ? DateTime.SpecifyKind(model.PublicationDate, DateTimeKind.Utc) : default;
-        }
-
-        public static DateTime GetNormalizedPublicationDate(this DateTime? publicationDate)
-        {
-            return publicationDate != null ? DateTime.SpecifyKind(publicationDate.Value, DateTimeKind.Utc) : default;
-        }
-
-        public static string GetDisplayNamePath(this InstanceModelBase model, List<NodeModel> nodesVisited)
-        {
-            if (nodesVisited.Contains(model))
-            {
-                return "(cycle)";
-            }
-            nodesVisited.Add(model);
-            if (model.Parent is InstanceModelBase parent)
-            {
-                return $"{parent.GetDisplayNamePath(nodesVisited)}.{model.DisplayName.FirstOrDefault()?.Text}";
-            }
-            return model.DisplayName.FirstOrDefault()?.Text;
-        }
-
-        public static string GetUnqualifiedBrowseName(this NodeModel nodeModel)
-        {
-            var browseName = nodeModel.GetBrowseName();
-            var parts = browseName.Split([';'], 2);
-
-            if (parts.Length > 1)
-            {
-                return parts[1];
-            }
-
-            return browseName;
-        }
-
-        internal static void SetEngineeringUnits(this VariableModel model, EUInformation euInfo)
-        {
-            model.EngineeringUnit = new VariableModel.EngineeringUnitInfo {
-                DisplayName = euInfo.DisplayName?.ToModelSingle(),
-                Description = euInfo.Description?.ToModelSingle(),
-                NamespaceUri = euInfo.NamespaceUri,
-                UnitId = euInfo.UnitId,
-            };
-        }
-
-        internal static void SetRange(this VariableModel model, Opc.Ua.Range euRange)
-        {
-            model.MinValue = euRange.Low;
-            model.MaxValue = euRange.High;
-        }
-
-        internal static void SetInstrumentRange(this VariableModel model, Opc.Ua.Range range)
-        {
-            model.InstrumentMinValue = range.Low;
-            model.InstrumentMaxValue = range.High;
-        }
-    }
-}
